Move the non-committing instVar e2e scenarios into the CI suite
`gci/gciInstVar.e2e.test.ts` had eight tests, five of which never commit, so
they belong in the suite that CI actually runs. Three of those five were already
duplicated by `refactoringInstVar.integration.test.ts`; the other two — the
shadowed-temporary prediction and the decline naming the offending subclass —
move there now, along with the selective copy-forward control (a method that
never referenced the removed variable must survive the new class version and
must not be reported as dropped) and the staged-change count on add.

What stays behind is only what the harness cannot host: the three tests that
COMMIT. `useIntegrationTest` aborts after every test and an abort cannot undo a
commit, so moving them needs a commit-and-compensate story for the harness —
a policy decision about its "never commit" invariant, tracked as its own phase
rather than forced here.
